Fix : cpp06, 07
- cpp06 ex00 strtof 는 11버전. strtod로 수정, 자료형도 float -> double
수정
- cpp07 ex02 대입연산자에서 leak발생. delete 추가

+++ cpp/cpp06/ex00/Conversion.hpp
@@ -6,7 +6,7 @@
 /*   By: daekim <daekim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/30 19:46:47 by daekim            #+#    #+#             */
-/*   Updated: 2021/11/30 19:47:33 by daekim           ###   ########.fr       */
+/*   Updated: 2021/12/12 19:22:26 by daekim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ class Conversion
 		int isprint;
 		int d_zero;
 		int f_zero;
-		float fnum;
+		double fnum;
 		double dounum;
 		char *argv;
